// src/types/generated/StoredAppSettingsPayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type StoredAppSettingsPayload = { importMode: string | null, libraryPath: string | null, loadRemoteImages: string | null, checkUpdatesOnStartup: string | null, externalBackupDir: string | null, };